ops/pipelines/airflow_dag.py
"""
Inscenium Video Processing Pipeline

This DAG orchestrates the complete video processing pipeline:
1. Ingest video content
2. Shot detection and segmentation
3. Computer vision analysis (SAM2, depth, flow)
4. Surface proposal generation
5. UAOR uncertainty fusion
6. SGI database updates
7. Render sidecar generation
8. Quality control validation
9. Asset packaging for edge delivery

Author: Inscenium Team
"""

from datetime import datetime, timedelta
from typing import Any, Dict

from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.operators.bash import BashOperator
from airflow.utils.dates import days_ago

# Default arguments for all tasks
default_args = {
    'owner': 'inscenium',
    'depends_on_past': False,
    'start_date': days_ago(1),
    'email_on_failure': True,
    'email_on_retry': False,
    'retries': 1,
    'retry_delay': timedelta(minutes=5),
}

# DAG definition
dag = DAG(
    'inscenium_video_pipeline',
    default_args=default_args,
    description='Inscenium video processing and SGI generation pipeline',
    schedule_interval=None,  # Triggered manually or via API
    catchup=False,
    max_active_runs=1,
    tags=['inscenium', 'computer-vision', 'sgi'],
)

def ingest_video(**context: Any) -> Dict[str, Any]:
    """
    Ingest video content and validate format.
    
    TODO: Implement actual video ingestion from various sources
    - S3 buckets
    - HTTP URLs  
    - Local filesystem
    - Streaming sources
    """
    import os
    from pathlib import Path
    
    # Get video path from DAG run configuration
    video_path = context['dag_run'].conf.get('video_path', 'tests/golden_scenes/assets/sample.mp4')
    title_id = context['dag_run'].conf.get('title_id', 1)
    
    print(f"Processing video: {video_path}")
    print(f"Title ID: {title_id}")
    
    # Validate video exists (stub implementation)
    if not Path(video_path).exists():
        print(f"WARNING: Video file not found: {video_path}")
        print("Using stub implementation for CI/development")
    
    # Return metadata for downstream tasks
    return {
        'video_path': video_path,
        'title_id': title_id,
        'duration_seconds': 120.0,
        'resolution': '1920x1080',
        'fps': 24.0
    }

def detect_shots(**context: Any) -> Dict[str, Any]:
    """
    Detect shot boundaries in video content.
    
    Uses PySceneDetect or similar shot detection algorithms.
    """
    import sys
    sys.path.append('.')
    
    from perception.shot_detect import detect_scenes
    
    # Get video metadata from upstream task
    video_metadata = context['task_instance'].xcom_pull(task_ids='ingest_video')
    video_path = video_metadata['video_path']
    
    print(f"Detecting shots in: {video_path}")
    
    # Run shot detection
    shots = detect_scenes(video_path)
    
    print(f"Detected {len(shots)} shots")
    for i, shot in enumerate(shots):
        print(f"  Shot {i+1}: {shot.start_time:.2f}s - {shot.end_time:.2f}s")
    
    return {
        'video_path': video_path,
        'shots': [shot.to_dict() for shot in shots],
        'shot_count': len(shots)
    }

def run_sam2_segmentation(**context: Any) -> Dict[str, Any]:
    """
    Run SAM2 segmentation on detected shots.
    
    Generates object masks and segmentation logits for each shot.
    """
    import sys
    sys.path.append('.')
    
    from perception.sam2_runner import run_sam2
    from perception.shot_detect import Shot
    
    # Get shot data from upstream task
    shot_data = context['task_instance'].xcom_pull(task_ids='detect_shots')
    video_path = shot_data['video_path']
    shots = [Shot.from_dict(s) for s in shot_data['shots']]
    
    print(f"Running SAM2 on {len(shots)} shots")
    
    # Run SAM2 segmentation
    masks, logits = run_sam2(video_path, shots)
    
    return {
        'video_path': video_path,
        'masks_path': masks,  # Path to saved masks
        'logits_path': logits,  # Path to saved logits
        'segmentation_complete': True
    }

def estimate_depth_flow(**context: Any) -> Dict[str, Any]:
    """
    Estimate depth maps and optical flow.
    
    Uses MiDaS for depth estimation and RAFT for optical flow.
    """
    import sys
    sys.path.append('.')
    
    from perception.depth_midas import estimate_depth
    from perception.flow_raft import estimate_flow
    from perception.shot_detect import Shot
    
    # Get data from upstream tasks
    shot_data = context['task_instance'].xcom_pull(task_ids='detect_shots')
    video_path = shot_data['video_path']
    shots = [Shot.from_dict(s) for s in shot_data['shots']]
    
    print("Estimating depth maps...")
    depth_maps, conf_maps = estimate_depth(video_path, shots)
    
    print("Estimating optical flow...")
    flows = estimate_flow(video_path, shots)
    
    return {
        'depth_maps_path': depth_maps,
        'confidence_maps_path': conf_maps,
        'flows_path': flows,
        'depth_flow_complete': True
    }

def generate_surface_proposals(**context: Any) -> Dict[str, Any]:
    """
    Generate surface placement proposals from computer vision outputs.
    
    Combines segmentation, depth, and flow to identify planar surfaces.
    """
    import sys
    sys.path.append('.')
    
    from perception.surfel_proposals import propose_surfaces
    
    # Get all upstream data
    seg_data = context['task_instance'].xcom_pull(task_ids='sam2_segmentation')
    depth_data = context['task_instance'].xcom_pull(task_ids='depth_flow_estimation')
    shot_data = context['task_instance'].xcom_pull(task_ids='detect_shots')
    
    print("Generating surface proposals...")
    
    # Combine all inputs for surface proposal
    proposals = propose_surfaces(
        masks_path=seg_data['masks_path'],
        depth_path=depth_data['depth_maps_path'],
        flows_path=depth_data['flows_path'],
        shots=[Shot.from_dict(s) for s in shot_data['shots']]
    )
    
    print(f"Generated {len(proposals)} surface proposals")
    
    return {
        'proposals': [p.to_dict() for p in proposals],
        'proposal_count': len(proposals)
    }

def run_uaor_fusion(**context: Any) -> Dict[str, Any]:
    """
    Run UAOR uncertainty-aware occlusion reasoning.
    
    Fuses multiple computer vision modalities with uncertainty estimates.
    """
    import sys
    sys.path.append('.')
    
    from perception.uaor_fuser import fuse_uaor
    
    print("Running UAOR fusion...")
    
    # Get all upstream outputs
    seg_data = context['task_instance'].xcom_pull(task_ids='sam2_segmentation')
    depth_data = context['task_instance'].xcom_pull(task_ids='depth_flow_estimation')
    proposal_data = context['task_instance'].xcom_pull(task_ids='surface_proposals')
    
    # Run UAOR fusion
    occlusion_maps, uncertainty_maps = fuse_uaor(
        masks_path=seg_data['masks_path'],
        depth_path=depth_data['depth_maps_path'],
        proposals=proposal_data['proposals']
    )
    
    return {
        'occlusion_maps_path': occlusion_maps,
        'uncertainty_maps_path': uncertainty_maps,
        'uaor_complete': True
    }

def update_sgi_database(**context: Any) -> Dict[str, Any]:
    """
    Update Scene Graph Intelligence database with analysis results.
    
    Persists shots, surfaces, and placement opportunities to PostgreSQL.
    """
    import sys
    sys.path.append('.')
    
    from sgi.sgi_builder import build_sgi
    
    print("Updating SGI database...")
    
    # Get all processing results
    video_data = context['task_instance'].xcom_pull(task_ids='ingest_video')
    shot_data = context['task_instance'].xcom_pull(task_ids='detect_shots')
    proposal_data = context['task_instance'].xcom_pull(task_ids='surface_proposals')
    uaor_data = context['task_instance'].xcom_pull(task_ids='uaor_fusion')
    
    # Build SGI entries
    sgi_id = build_sgi(
        title_id=video_data['title_id'],
        shots=shot_data['shots'],
        proposals=proposal_data['proposals'],
        occlusion_maps=uaor_data['occlusion_maps_path'],
        uncertainty_maps=uaor_data['uncertainty_maps_path']
    )
    
    print(f"SGI database updated, entry ID: {sgi_id}")
    
    return {
        'sgi_id': sgi_id,
        'database_updated': True
    }

def generate_render_assets(**context: Any) -> Dict[str, Any]:
    """
    Generate rendering assets and HLS sidecars.
    
    Creates alpha masks, depth maps, normal maps, and spherical harmonics probes.
    """
    import sys
    sys.path.append('.')
    
    from render.sidecar_packager import package_sidecar
    
    print("Generating render assets...")
    
    # Get processing outputs
    video_data = context['task_instance'].xcom_pull(task_ids='ingest_video')
    depth_data = context['task_instance'].xcom_pull(task_ids='depth_flow_estimation')
    uaor_data = context['task_instance'].xcom_pull(task_ids='uaor_fusion')
    
    # Package HLS sidecars
    sidecar_manifest = package_sidecar(
        video_path=video_data['video_path'],
        depth_maps=depth_data['depth_maps_path'],
        occlusion_maps=uaor_data['occlusion_maps_path'],
        uncertainty_maps=uaor_data['uncertainty_maps_path']
    )
    
    print(f"Generated sidecar manifest: {sidecar_manifest}")
    
    return {
        'sidecar_manifest': sidecar_manifest,
        'assets_generated': True
    }

def run_quality_control(**context: Any) -> Dict[str, Any]:
    """
    Run quality control validation on processing results.
    
    Calculates PRS scores, validates asset integrity, checks thresholds.
    """
    import sys
    sys.path.append('.')
    
    from render.qc_metrics import calculate_prs_score
    from measure.prs_meter import meter_prs
    
    print("Running quality control...")
    
    # Get render assets
    render_data = context['task_instance'].xcom_pull(task_ids='render_assets')
    video_data = context['task_instance'].xcom_pull(task_ids='ingest_video')
    
    # Calculate PRS score
    prs_score = calculate_prs_score(
        video_path=video_data['video_path'],
        sidecar_manifest=render_data['sidecar_manifest']
    )
    
    print(f"PRS Score: {prs_score:.2f}")
    
    # Check quality thresholds
    quality_passed = prs_score >= 70.0  # Minimum threshold
    
    if not quality_passed:
        print(f"WARNING: PRS score {prs_score:.2f} below threshold")
    
    return {
        'prs_score': prs_score,
        'quality_passed': quality_passed,
        'qc_complete': True
    }

# Define task dependencies
ingest_task = PythonOperator(
    task_id='ingest_video',
    python_callable=ingest_video,
    dag=dag,
)

shot_detection_task = PythonOperator(
    task_id='detect_shots',
    python_callable=detect_shots,
    dag=dag,
)

sam2_task = PythonOperator(
    task_id='sam2_segmentation',
    python_callable=run_sam2_segmentation,
    dag=dag,
)

depth_flow_task = PythonOperator(
    task_id='depth_flow_estimation',
    python_callable=estimate_depth_flow,
    dag=dag,
)

surface_task = PythonOperator(
    task_id='surface_proposals',
    python_callable=generate_surface_proposals,
    dag=dag,
)

uaor_task = PythonOperator(
    task_id='uaor_fusion',
    python_callable=run_uaor_fusion,
    dag=dag,
)

sgi_task = PythonOperator(
    task_id='update_sgi',
    python_callable=update_sgi_database,
    dag=dag,
)

render_task = PythonOperator(
    task_id='render_assets',
    python_callable=generate_render_assets,
    dag=dag,
)

qc_task = PythonOperator(
    task_id='quality_control',
    python_callable=run_quality_control,
    dag=dag,
)

# Set up task dependencies
ingest_task >> shot_detection_task
shot_detection_task >> [sam2_task, depth_flow_task]
[sam2_task, depth_flow_task] >> surface_task
[surface_task, sam2_task, depth_flow_task] >> uaor_task
uaor_task >> sgi_task
[uaor_task, depth_flow_task] >> render_task
render_task >> qc_task