perception/depth_midas.py
"""
MiDaS Depth Estimation for Inscenium
====================================

Monocular depth estimation using MiDaS models for scene understanding
and surface depth analysis in placement opportunities.
"""

import logging
import numpy as np
from typing import Optional, Tuple, Dict, Any
import cv2
from pathlib import Path

logger = logging.getLogger(__name__)

class MiDaSDepthEstimator:
    """MiDaS-based depth estimation for placement analysis"""
    
    def __init__(self, model_type: str = "DPT_Large", device: str = "cpu"):
        self.model_type = model_type
        self.device = device
        self.model = None
        self.transform = None
        
    def initialize(self) -> bool:
        """Initialize MiDaS model and transforms"""
        try:
            # Mock initialization for development
            logger.info(f"Initializing MiDaS {self.model_type} on {self.device}")
            return True
        except Exception as e:
            logger.error(f"Failed to initialize MiDaS model: {e}")
            return False
    
    def estimate_depth(self, image: np.ndarray) -> Optional[np.ndarray]:
        """
        Estimate depth map from input image
        
        Args:
            image: Input RGB image (H, W, 3)
            
        Returns:
            Depth map (H, W) with normalized depth values
        """
        if image is None or len(image.shape) != 3:
            logger.error("Invalid input image for depth estimation")
            return None
            
        try:
            # Mock depth estimation - generate realistic depth map
            height, width = image.shape[:2]
            
            # Create synthetic depth based on image gradients
            gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
            
            # Simple depth estimation based on intensity and gradients
            depth = self._generate_mock_depth(gray)
            
            # Normalize to 0-1 range
            depth_normalized = (depth - depth.min()) / (depth.max() - depth.min())
            
            logger.debug(f"Generated depth map: {depth_normalized.shape}")
            return depth_normalized.astype(np.float32)
            
        except Exception as e:
            logger.error(f"Depth estimation failed: {e}")
            return None
    
    def _generate_mock_depth(self, gray_image: np.ndarray) -> np.ndarray:
        """Generate mock depth map for development/testing"""
        height, width = gray_image.shape
        
        # Create depth based on multiple cues
        # 1. Center bias (objects in center tend to be closer)
        y_coords, x_coords = np.ogrid[:height, :width]
        center_y, center_x = height // 2, width // 2
        center_distance = np.sqrt((x_coords - center_x)**2 + (y_coords - center_y)**2)
        center_depth = 1.0 - (center_distance / np.max(center_distance))
        
        # 2. Vertical bias (lower regions tend to be closer)
        vertical_depth = 1.0 - (y_coords / height)
        
        # 3. Intensity bias (brighter objects often closer)
        intensity_depth = gray_image.astype(np.float32) / 255.0
        
        # 4. Edge information (edges often indicate depth discontinuities)
        edges = cv2.Canny(gray_image, 50, 150)
        edge_depth = cv2.GaussianBlur(edges.astype(np.float32), (5, 5), 1.0) / 255.0
        
        # Combine depth cues
        depth = (
            center_depth * 0.3 +
            vertical_depth * 0.3 +
            intensity_depth * 0.2 +
            edge_depth * 0.2
        )
        
        # Add some noise for realism
        noise = np.random.normal(0, 0.05, depth.shape)
        depth = np.clip(depth + noise, 0, 1)
        
        # Smooth the depth map
        depth = cv2.GaussianBlur(depth, (7, 7), 1.5)
        
        return depth
    
    def get_depth_at_points(self, depth_map: np.ndarray, points: np.ndarray) -> np.ndarray:
        """
        Get depth values at specific pixel coordinates
        
        Args:
            depth_map: Depth map (H, W)
            points: Array of (x, y) coordinates, shape (N, 2)
            
        Returns:
            Array of depth values at specified points
        """
        if depth_map is None or points is None:
            return np.array([])
            
        try:
            height, width = depth_map.shape
            depths = []
            
            for point in points:
                x, y = int(point[0]), int(point[1])
                if 0 <= x < width and 0 <= y < height:
                    depths.append(depth_map[y, x])
                else:
                    depths.append(0.0)  # Default depth for out-of-bounds points
            
            return np.array(depths)
            
        except Exception as e:
            logger.error(f"Failed to get depth at points: {e}")
            return np.array([])
    
    def analyze_surface_depth(self, depth_map: np.ndarray, mask: np.ndarray) -> Dict[str, Any]:
        """
        Analyze depth characteristics of a surface region
        
        Args:
            depth_map: Depth map (H, W)
            mask: Binary mask of surface region (H, W)
            
        Returns:
            Dictionary with depth analysis results
        """
        if depth_map is None or mask is None:
            return {}
            
        try:
            surface_depths = depth_map[mask > 0]
            
            if len(surface_depths) == 0:
                return {"error": "Empty surface region"}
            
            analysis = {
                "mean_depth": float(np.mean(surface_depths)),
                "median_depth": float(np.median(surface_depths)),
                "depth_std": float(np.std(surface_depths)),
                "min_depth": float(np.min(surface_depths)),
                "max_depth": float(np.max(surface_depths)),
                "depth_range": float(np.max(surface_depths) - np.min(surface_depths)),
                "surface_flatness": self._calculate_flatness(surface_depths),
                "depth_confidence": self._calculate_depth_confidence(surface_depths)
            }
            
            return analysis
            
        except Exception as e:
            logger.error(f"Surface depth analysis failed: {e}")
            return {"error": str(e)}
    
    def _calculate_flatness(self, depths: np.ndarray) -> float:
        """Calculate surface flatness score (0-1, higher = flatter)"""
        if len(depths) < 2:
            return 0.0
        
        # Use coefficient of variation as flatness measure
        cv = np.std(depths) / (np.mean(depths) + 1e-8)
        flatness = max(0.0, 1.0 - cv * 5.0)  # Scale and invert
        return float(flatness)
    
    def _calculate_depth_confidence(self, depths: np.ndarray) -> float:
        """Calculate confidence in depth estimation (0-1)"""
        if len(depths) < 5:
            return 0.5  # Low confidence for small regions
        
        # Higher confidence for consistent depth values
        consistency = 1.0 - (np.std(depths) / (np.mean(depths) + 1e-8))
        confidence = np.clip(consistency, 0.0, 1.0)
        return float(confidence)

# Mock depth estimation function for testing
def mock_depth_estimation(image: np.ndarray) -> Dict[str, Any]:
    """Generate mock depth analysis for CI testing"""
    height, width = image.shape[:2] if len(image.shape) >= 2 else (480, 640)
    
    return {
        "depth_map_shape": (height, width),
        "mean_depth": 0.623,
        "depth_range": 0.845,
        "surface_count": 5,
        "flat_surfaces": 3,
        "analysis": {
            "foreground_depth": 0.8,
            "background_depth": 0.2,
            "depth_layers": 3,
            "occlusion_boundaries": 12
        }
    }

if __name__ == "__main__":
    # Demo usage
    estimator = MiDaSDepthEstimator()
    if estimator.initialize():
        # Generate mock image
        test_image = np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8)
        depth_map = estimator.estimate_depth(test_image)
        
        if depth_map is not None:
            print(f"Depth map generated: {depth_map.shape}")
            print(f"Depth range: {depth_map.min():.3f} - {depth_map.max():.3f}")
        else:
            print("Depth estimation failed")
    else:
        print("Model initialization failed")