pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
name = "inscenium"
version = "1.0.0"
description = "Advanced computer vision and rendering system for contextual ad placement"
authors = ["Inscenium Contributors <dev@inscenium.dev>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/inscenium/inscenium"
repository = "https://github.com/inscenium/inscenium"
documentation = "https://docs.inscenium.dev"
keywords = ["computer-vision", "advertising", "webassembly", "cuda", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[[tool.poetry.source]]
name = "inscenium"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python-headless = "4.8.1.78"
numpy = "<2.0"
scipy = "1.11.4"
scikit-image = "0.22.0"
onnxruntime = "1.16.3"
fastapi = {version = "0.104.1", optional = true}
uvicorn = {version = "0.24.0", optional = true}
psycopg2-binary = "2.9.9"
sqlalchemy = "2.0.23"
alembic = "1.13.1"
click = "8.1.7"
rich = "13.7.0"
pyjwt = "2.8.0"
httpx = "0.25.2"
redis = "5.0.1"
kafka-python = "2.0.2"
opentelemetry-sdk = "1.21.0"
opentelemetry-exporter-otlp = "1.21.0"
opentelemetry-instrumentation-fastapi = "0.42b0"
opentelemetry-instrumentation-psycopg2 = "0.42b0"
lpips = "0.1.4"
matplotlib = "3.8.2"
seaborn = "0.13.0"
tqdm = "4.66.1"
structlog = "23.2.0"
typer = "0.9.0"
jinja2 = "3.1.2"
torch = "2.2.2"
torchvision = "0.17.2"
pillow = ">=10.2.0"
pydantic = ">=2.7"

[tool.poetry.extras]
web = ["fastapi", "uvicorn"]

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-cov = "4.1.0"
pytest-asyncio = "0.21.1"
pytest-mock = "3.12.0"
black = "23.12.1"
isort = "5.13.2"
ruff = "0.1.8"
mypy = "1.8.0"
types-requests = "2.31.0.10"
types-redis = "4.6.0.11"
types-pillow = "10.1.0.2"
pre-commit = "3.6.0"
sphinx = "7.2.6"
sphinx-rtd-theme = "2.0.0"
jupyter = "1.0.0"
ipykernel = "6.27.1"

[tool.poetry.scripts]
inscenium = "inscenium.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Exclude directories
  \.eggs
  | \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "ICN",  # flake8-import-conventions
    "S",    # bandit security
]
ignore = [
    "S101",  # assert usage (common in tests)
    "S603",  # subprocess call check
    "S607",  # subprocess partial executable path
]
exclude = [
    ".eggs",
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ARG001", "ARG002"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "cv2.*",
    "lpips.*",
    "kafka.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests as requiring GPU",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["perception", "sgi", "render", "edge", "measure"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]